requirements.txt
asgiref==3.8.1
Django==5.0.1
django-ckeditor==6.7.1
django-cleanup==9.0.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
pillow==10.4.0
sqlparse==0.5.1
